+++ src/osha/oira/upgrade/v1/20240214105028_index_manager_roles/upgrade.py
@@ -1,0 +1,12 @@
+from ftw.upgrade import UpgradeStep
+from plone import api
+
+
+class IndexManagerRoles(UpgradeStep):
+    """Index manager roles."""
+
+    def __call__(self):
+        self.install_upgrade_profile()
+        client = api.portal.get().client
+        for brain in api.content.find(context=client, depth=3):
+            brain.getObject().reindexObject(idxs=["managerRolesAndUsers"])